+++ src/parser/inline/math.ts
@@ -10,12 +10,12 @@ export const cache = new Cache<string, HTMLElement>(20); // for rerendering in e
 export const math: MathParser = subline(verify(
   fmap(
     stringify(
-      surround('$', compress(some(union([escsource]), /^[\n$]/)), /^\$(?![0-9])/)),
+      surround('${', compress(some(union([escsource]), /^}\$|^\n/)), /^}\$/)),
     ([body]) => {
-      const source = `$${body}$`; // TODO: Should use String.prototype.trimEnd.
+      const source = `$\{${body}}$`; // TODO: Should use String.prototype.trimEnd.
       if (cache.has(source)) return [cache.get(source)!.cloneNode(true)];
       const el = html('span', { class: 'math notranslate' }, source);
       void el.setAttribute('data-src', source);
       return [el];
     }),
-  ([el]) => startsWithTightText(frag(el.textContent!.slice(1, -1)))));
+  ([el]) => startsWithTightText(frag(el.textContent!.slice(2, -2)))));